Add a simple "VAD" alternative - periodic VAD
This is a very simple alternative to using Silero Vad, where every
5 minutes (0:00 - 5:00, 5:00 - 10:00, etc.) is marked as a
"speech" segment, and whisper is then run on each
segment individually.

The upside is that all potential dialogue will be sent to
Whisper for transcribing, but it is also more likely to
get into an infinite sentence loop (although limited to each
5 minute segment).

Creating an artifical break every 5 minutes may also break
up a sentence accidentally, causing the sentence (or a word) to be
incorrectly transcribed.